Add Style Edit Menu

+++ src/app/src/main/java/kaleidot725/highestpeaks/di/data/Style.kt
@@ -1,3 +1,3 @@
 package kaleidot725.highestpeaks.di.data
 
-data class Style(val dp : Int, val detail : String)
+data class Style(val dp : Float, val detail : String)
